+++ source/Public/Assert-BoundParameter.ps1
@@ -5,7 +5,7 @@
 
     .DESCRIPTION
         This command asserts passed parameters. It takes a hashtable, normally
-        `$PSBoundParameters`. There are two parameter sets for this command.
+        `$PSBoundParameters`. There are three parameter sets for this command.
 
         >There is no built in logic to validate against parameters sets for DSC
         >so this can be used instead to validate the parameters that were set in
@@ -22,6 +22,11 @@
         Assert that required parameters has been specified, and throws an exception
         if not. Optionally it can be specified that parameters are only required
         if a specific parameter has been passed.
+
+        **AtLeastOne**
+
+        Assert that at least one parameter from the specified list has been bound,
+        and throws an exception if none are present.
 
     .PARAMETER BoundParameterList
         The parameters that should be evaluated against the mutually exclusive
@@ -46,6 +51,9 @@
        One or more parameter names that if specified will trigger the evaluation.
        If neither of the parameter names has been specified the evaluation of required
        parameters are not made.
+
+    .PARAMETER AtLeastOneList
+       An array of parameter names where at least one must be bound.
 
     .EXAMPLE
         $assertBoundParameterParameters = @{
@@ -77,6 +85,11 @@
         Assert-BoundParameter -BoundParameterList $PSBoundParameters -RequiredParameter @('PBStartPortRange', 'PBEndPortRange') -RequiredBehavior 'AtLeastOnce'
 
         Throws an exception if at least one of the two parameters are not specified.
+
+    .EXAMPLE
+        Assert-BoundParameter -BoundParameterList $PSBoundParameters -AtLeastOneList @('Severity', 'MessageId')
+
+        Throws an exception if none of the parameters 'Severity' or 'MessageId' are specified.
 #>
 function Assert-BoundParameter
 {
@@ -106,7 +119,11 @@ function Assert-BoundParameter
 
         [Parameter(ParameterSetName = 'RequiredParameter')]
         [System.String[]]
-        $IfParameterPresent
+        $IfParameterPresent,
+
+        [Parameter(ParameterSetName = 'AtLeastOne', Mandatory = $true)]
+        [System.String[]]
+        $AtLeastOneList
     )
 
     switch ($PSCmdlet.ParameterSetName)
@@ -122,7 +139,7 @@ function Assert-BoundParameter
                     $script:localizedData.ParameterUsageWrong `
                     -f ($MutuallyExclusiveList1 -join "','"), ($MutuallyExclusiveList2 -join "','")
 
-                New-InvalidArgumentException -ArgumentName 'Parameters' -Message $errorMessage
+                New-ArgumentException -ArgumentName 'Parameters' -Message $errorMessage
             }
 
             break
@@ -136,6 +153,22 @@ function Assert-BoundParameter
             }
 
             Assert-RequiredCommandParameter @PSBoundParameters
+            break
+        }
+
+        'AtLeastOne'
+        {
+            $boundParametersFromList = $BoundParameterList.Keys.Where({ $_ -in $AtLeastOneList })
+
+            if ($boundParametersFromList.Count -eq 0)
+            {
+                $errorMessage = `
+                    $script:localizedData.Assert_BoundParameter_AtLeastOneParameterMustBeSet `
+                    -f ($AtLeastOneList -join "','")
+
+                New-ArgumentException -ArgumentName 'Parameters' -Message $errorMessage
+            }
+
             break
         }
     }